Move /volumes to top level feature.
